pyproject.toml
[project]
name = "pytest-projet"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "allure-pytest>=2.15.0",
    "allure-python-commons>=2.15.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
